notification-service/requirements.txt
Flask==3.0.0
kafka-python==2.0.2
python-dotenv==1.0.0
Werkzeug==3.0.1
